Support the nested settings dictionaries in ES 1.x.
Starting with Elasticsearch 1.0 (probably), the previously
flat settings dictionaries returned by the /_settings and
/_cluster/settings endpoints became nested, i.e. instead of
getting back

    {
      "transient": {
        "cluster.routing.allocation.disable_allocation": "false"
      }
    }

we get

    {
      "transient": {
        "cluster": {
          "routing": {
            "allocation": {
              "disable_allocation": "false"
            }
          }
        }
      }
    }

which causes us to not recognize what settings we have, leading to
a failing assertion. We now try both ways of accessing the settings
dictionaries to remain compatible with both Elasticsearch 0.x and 1.x.

Closes bug #4.